GS-Signal

A tbl containing the coverage from every human sample in RLBase around "gold-standard" R-loop sites.

gs_signal(quiet = FALSE)

Arguments

quiet
If TRUE, messages are suppressed. Default: FALSE.

Value

A tbl object. See details.

Details

Source

Gold-standard R-loop sites are those which were profiled with ultra-long-read R-loop sequencing (SMRF-Seq) as described in Malig et al., 2021.. A bed file containing these sites was derived and lifted to hg38. Sites were then extended by 100kb bi-directionally and then binned into 1kb bins.

For each sample in RLBase, coverage tracks were calculated from alignments using RLPipes. Then, coverage tracks were summed within each of the 1kb bins using deepTools and wrangled into a tbl.

The processing steps are described in the RLBase-data README.md.

These data are used by RLSeq::corrAnalyze() to determine how well a query sample correlates with the samples in RLBase.

Structure

Sample of data:

location	ERX2277510	ERX2277511	ERX3974959	...

chr1_67636071_67637071	23	22	93	...

chr1_67637071_67638071	44	39	84	...

chr1_67638071_67639071	26	30	100	...

...	...	...	...	...

Column desc:

location - 1kb bins within "gold-standard" R-loop regions profiled with ultra-long-read R-loop sequencing (SMRF-Seq).

ERX.../SRX... - Columns corresponding to RLBase samples (see rlbase_samples). For each sample, the total coverage (bigWig file) within the range is shown.

Examples

gss <- gs_signal()