R-loop Binding Proteins

A list of annotations for genomic features relevant to R-loops. Both mm10 and hg38 annotations are available.

rlbps(quiet = FALSE)

Arguments

quiet
If TRUE, messages are suppressed. Default: FALSE.

Value

A tbl object. See details.

Details

Source

R-Loop binding proteins (RLBPs) are proteins which were discovered by mass-spectrometry after S9.6 immuno-precipitation (S9.6 is an antibody that binds R-loops). The data were aggregated from three studies which analyzed RLBPs in this manner:

Nadel et al., 2015 - article

Wang et al., 2018 - article

Cristini et al., 2018 - article

The data were downloaded from the supplementary materials of each study and processed using custom R scripts as part of the RLBase-data pipeline. This processing involved normalization of mass-spec abundance calculations and the derivation of a "combined score" for ranking all the proteins.

Structure

geneName	NADEL.2015	WANG.2018	CRISTINI.2018	combinedScore

DDX21	0.9	0	0.968	1.87

DHX9	0.85	0.0635	0.831	1.74

...	...	...	...	...

Columns:

geneName

The official HGNC symbol of the RLBP

NADEL.2015

Normalized abundance of protein from Nadel et al., 2015

WANG.2018

Normalized abundance of protein from Wang et al., 2018

CRISTINI.2018

Normalized abundance of protein from Cristini et al., 2018

combinedScore

Sum of normalized abundance across studies.

Examples

rlbpsData <- rlbps()